🚀 [Feat]: auth 디렉토리 생성
