+++ src/app/(auth)/create-study/components/ContentAdderBox.tsx
