+++ src/app/(auth)/create-study/components/ContentInputBox.tsx
