+++ src/app/(auth)/create-study/components/LabelEditor.tsx
