+++ src/app/(auth)/create-study/components/StudyApply.tsx
