+++ src/app/(auth)/create-study/components/StudyApplyQuestion.tsx
