+++ src/app/(auth)/create-study/components/StudyBasicInfo.tsx
