+++ src/app/(auth)/create-study/components/StudyBenefits.tsx
