+++ src/app/(auth)/create-study/components/StudyCurriculum.tsx
