+++ src/app/(auth)/create-study/components/StudyDescription.tsx
