+++ src/app/(auth)/create-study/components/StudyMethod.tsx
