+++ src/app/(auth)/create-study/components/StudyRules.tsx
