+++ src/app/(auth)/create-study/layout.tsx
