+++ src/app/(auth)/mypage/components/ActiveStudies.tsx
