+++ src/app/(auth)/mypage/components/Statistics.tsx
