+++ src/app/(auth)/mypage/components/StudySchedule.tsx
