+++ src/app/(auth)/study-room/[id]/archive/page.tsx
@@ -1,4 +1,4 @@
-import ArchiveCard from "@/app/study-room/components/archive/ArchiveCard";
+import ArchiveCard from "@/app/(auth)/study-room/components/archive/ArchiveCard";
 
 const ArchivePage = () => {
   return <ArchiveCard />;
