+++ src/app/(auth)/study-room/[id]/curriculum/page.tsx
@@ -1,0 +1,7 @@
+import CurriculumCard from "@/app/(auth)/study-room/components/curriculum/CurriculumCard";
+
+const Curriculum = () => {
+  return <CurriculumCard />;
+};
+
+export default Curriculum;
