+++ src/app/(auth)/study-room/[id]/member/layout.tsx
