+++ src/app/(auth)/study-room/[id]/notice/page.tsx
@@ -1,0 +1,7 @@
+import NoticeCard from "@/app/(auth)/study-room/components/notice/NoticeCard";
+
+const Notice = () => {
+  return <NoticeCard />;
+};
+
+export default Notice;
