+++ src/app/(auth)/study-room/[id]/overview/page.tsx
