+++ src/app/(auth)/study-room/components/archive/ArchiveCard.tsx
