+++ src/app/(auth)/study-room/components/chat/ChatCard.tsx
