+++ src/app/(auth)/study-room/components/curriculum/Curriculum.tsx
