+++ src/app/(auth)/study-room/components/curriculum/CurriculumCard.tsx
