+++ src/app/(auth)/study-room/components/layout/StudyRoomIntendedCard.tsx
