+++ src/app/(auth)/study-room/components/layout/StudyRoomSessionCard.tsx
