+++ src/app/(auth)/study-room/components/layout/StudyRoomSideBarCard.tsx
