+++ src/app/(auth)/study-room/components/layout/StudyRoomTitleCard.tsx
