+++ src/app/(auth)/study-room/components/manage-applicants/InfoModal.tsx
