+++ src/app/(auth)/study-room/components/manage-applicants/ManageCard.tsx
