+++ src/app/(auth)/study-room/components/manage-overview/ContentInputBox.tsx
