+++ src/app/(auth)/study-room/components/manage-overview/PreviewBox.tsx
