+++ src/app/(auth)/study-room/components/member/MemberCard.tsx
