+++ src/app/(auth)/study-room/components/notice/ImportantNoticeBar.tsx
