+++ src/app/(auth)/study-room/components/notice/NoticeCard.tsx
