+++ src/app/(auth)/study-room/components/notice/NoticeModal.tsx
