--- src/app/study-room/[id]/curriculum/page.tsx
@@ -1,7 +1,0 @@
-import CurriculumCard from "@/app/study-room/components/curriculum/CurriculumCard";
-
-const Curriculum = () => {
-  return <CurriculumCard />;
-};
-
-export default Curriculum;
