--- src/app/study-room/[id]/notice/page.tsx
@@ -1,7 +1,0 @@
-import NoticeCard from "@/app/study-room/components/notice/NoticeCard";
-
-const Notice = () => {
-  return <NoticeCard />;
-};
-
-export default Notice;
