+++ src/features/create-study/form/CreateStudyForm1.tsx
@@ -6,9 +6,9 @@ import Typography from "@/shared/components/atoms/Typography";
 import ActionConfirmModal from "@/shared/components/molecules/ActionConfirmModal";
 import useModal from "@/shared/hooks/useModal";
 
-import StudyActions from "@/app/create-study/components/StudyActions";
-import StudyBasicInfo from "@/app/create-study/components/StudyBasicInfo";
-import StudyMethod from "@/app/create-study/components/StudyMethod";
+import StudyActions from "@/app/(auth)/create-study/components/StudyActions";
+import StudyBasicInfo from "@/app/(auth)/create-study/components/StudyBasicInfo";
+import StudyMethod from "@/app/(auth)/create-study/components/StudyMethod";
 import { StudyFormInterface } from "../types/create-study.type";
 
 const CreateStudyForm1 = () => {
