+++ src/features/create-study/form/CreateStudyForm2.tsx
@@ -5,10 +5,10 @@ import Badge from "@/shared/components/atoms/Badge";
 import Typography from "@/shared/components/atoms/Typography";
 import URL from "@/shared/constants/URL";
 
-import StudyActions from "@/app/create-study/components/StudyActions";
-import StudyBenefits from "@/app/create-study/components/StudyBenefits";
-import StudyDescription from "@/app/create-study/components/StudyDescription";
-import StudyRules from "@/app/create-study/components/StudyRules";
+import StudyActions from "@/app/(auth)/create-study/components/StudyActions";
+import StudyBenefits from "@/app/(auth)/create-study/components/StudyBenefits";
+import StudyDescription from "@/app/(auth)/create-study/components/StudyDescription";
+import StudyRules from "@/app/(auth)/create-study/components/StudyRules";
 import { StudyFormInterface } from "../types/create-study.type";
 
 const CreateStudyForm2 = () => {
